Add a switch_logical module
