+++ linkerd/app/outbound/src/lib.rs
@@ -11,14 +11,14 @@ pub mod http;
 mod ingress;
 pub mod logical;
 mod resolve;
+mod switch_logical;
 pub mod tcp;
 #[cfg(test)]
 pub(crate) mod test_util;
 
-use self::{endpoint::Endpoint, logical::Logical};
 use linkerd_app_core::{
     config::{ProxyConfig, ServerConfig},
-    io, metrics, profiles,
+    metrics, profiles,
     proxy::{
         api_resolve::{ConcreteAddr, Metadata},
         core::Resolve,
@@ -27,9 +27,9 @@ use linkerd_app_core::{
     svc::{self, stack::Param},
     tls,
     transport::{self, addrs::*, listen::Bind},
-    AddrMatch, Conditional, Error, Never, ProxyRuntime,
+    AddrMatch, Conditional, Error, ProxyRuntime,
 };
-use std::{collections::HashMap, fmt, future::Future, time::Duration};
+use std::{collections::HashMap, future::Future, time::Duration};
 use tracing::info;
 
 const EWMA_DEFAULT_RTT: Duration = Duration::from_millis(30);
@@ -109,76 +109,6 @@ impl<S> Outbound<S> {
             config: self.config,
             runtime: self.runtime,
             stack: self.stack.push(layer),
-        }
-    }
-
-    pub fn push_switch_logical<T, I, N, NSvc, SSvc>(
-        self,
-        logical: N,
-    ) -> Outbound<
-        impl svc::NewService<
-                (Option<profiles::Receiver>, T),
-                Service = impl svc::Service<I, Response = (), Error = Error, Future = impl Send>,
-            > + Clone,
-    >
-    where
-        Self: Clone + 'static,
-        T: Param<OrigDstAddr> + Clone + Send + Sync + 'static,
-        I: io::AsyncRead + io::AsyncWrite + io::PeerAddr + fmt::Debug + Send + Unpin + 'static,
-        N: svc::NewService<tcp::Logical, Service = NSvc> + Clone,
-        NSvc: svc::Service<I, Response = (), Error = Error>,
-        NSvc::Future: Send,
-        S: svc::NewService<tcp::Endpoint, Service = SSvc> + Clone,
-        SSvc: svc::Service<I, Response = (), Error = Error>,
-        SSvc::Future: Send,
-    {
-        let no_tls_reason = self.no_tls_reason();
-        let Self {
-            config,
-            runtime,
-            stack: endpoint,
-        } = self;
-
-        let stack = endpoint.push_switch(
-            move |(profile, target): (Option<profiles::Receiver>, T)| -> Result<_, Never> {
-                if let Some(rx) = profile {
-                    let profiles::Profile {
-                        ref addr,
-                        ref endpoint,
-                        opaque_protocol,
-                        ..
-                    } = *rx.borrow();
-
-                    if let Some((addr, metadata)) = endpoint.clone() {
-                        return Ok(svc::Either::A(Endpoint::from_metadata(
-                            addr,
-                            metadata,
-                            no_tls_reason,
-                            opaque_protocol,
-                        )));
-                    }
-
-                    if let Some(logical_addr) = addr.clone() {
-                        return Ok(svc::Either::B(Logical::new(
-                            logical_addr,
-                            rx.clone(),
-                            target.param(),
-                        )));
-                    }
-                }
-
-                Ok(svc::Either::A(Endpoint::forward(
-                    target.param(),
-                    no_tls_reason,
-                )))
-            },
-            logical,
-        );
-
-        Outbound {
-            config,
-            runtime,
-            stack,
         }
     }
 }
